add/update test

+++ open-care-backend/src/main/java/com/ciphertext/opencarebackend/repository/DoctorWorkplaceRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
  * @author Sadman
  */
 @Repository
-public interface DoctorWorkplaceRepository extends JpaRepository<DoctorWorkplace, Long> {
+public interface DoctorWorkplaceRepository extends JpaRepository<DoctorWorkplace, Integer> {
 }
